Refactor Auth permission principals to entries
